# public/openapi.yaml
openapi: 3.1.0
info:
  title: ChronoChimp API
  version: v1.0.0
  description: API for managing tasks and user settings in the ChronoChimp application.
servers:
  - url: /api
    description: Main API server
components:
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized: Invalid or missing authentication token"
              error:
                type: string
                example: "unauthorized"
              statusCode:
                type: integer
                example: 401
...
    ValidationError:
      description: Bad Request - Invalid input data.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Internal Server Error - E.g., JWT_SECRET not configured.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Internal Server Error"
  schemas:
...
    UserRole:
      type: string
      enum: [Admin, Editor, Viewer]
      example: Viewer
    User: # For general user display, not for sensitive creation/update
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "user-abc-123"
        name:
          type: string
          example: "Jane Doe"
        email:
          type: string
          format: email
          example: "jane.doe@example.com"
        role:
          $ref: '#/components/schemas/UserRole'
        avatarUrl:
          type: string
          format: uri
          nullable: true
          example: "https://placehold.co/100x100.png"
        isTwoFactorEnabled:
          type: boolean
          description: "Indicates if 2FA is enabled for the user."
          example: false
        emailNotificationsEnabled:
          type: boolean
          description: "User preference for email notifications."
          example: true
        inAppNotificationsEnabled:
          type: boolean
          description: "User preference for in-app notifications."
          example: true
        smtpHost:
          type: string
          nullable: true
          example: "smtp.example.com"
        smtpPort:
          type: integer
          nullable: true
          example: 587
        smtpEncryption:
          type: string
          enum: [none, ssl, starttls]
          nullable: true
          example: "starttls"
        smtpUsername:
          type: string
          nullable: true
          example: "user@example.com"
        smtpSendFrom:
          type: string
          format: email
          nullable: true
          description: "Default 'send from' email for SMTP."
          example: "noreply@example.com"
        joinedDate:
          type: string
          format: date-time
          example: "2023-01-01T12:00:00Z"
    UserProfile: # Schema for /api/me/profile
      allOf:
        - $ref: '#/components/schemas/User'
        # UserProfile might have slightly different fields or emphasis
        # For now, it's the same as User for GET, but PUT is different
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          example: "Jane Doe Smith"
        email:
          type: string
          format: email
          example: "jane.smith@example.com"
        avatarUrl:
          type: string
          format: uri # Or allow data URI: pattern: "^(https?://|data:image/)"
          description: "URL or Data URI for the user's avatar."
          nullable: true
          example: "https://placehold.co/100x100.png"
        isTwoFactorEnabled:
          type: boolean
        emailNotificationsEnabled:
          type: boolean
        inAppNotificationsEnabled:
          type: boolean
        smtpHost:
          type: string
          nullable: true
        smtpPort:
          type: integer
          nullable: true
        smtpEncryption:
          type: string
          enum: [none, ssl, starttls]
          nullable: true
        smtpUsername:
          type: string
          nullable: true
        smtpPassword:
          type: string
          format: password
          nullable: true
          description: "User's SMTP password. Sent only on update, not returned."
        smtpSendFrom:
          type: string
          format: email
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid input."
        details: # For Zod validation errors
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email: ["Invalid email address."]
            password: ["Password must be at least 6 characters."]
...
  securitySchemes:
    bearerAuth: # Can be used to denote JWT authentication
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /me/profile:
    get:
      summary: Get current user's profile
      tags: [User Settings]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      summary: Update current user's profile
      tags: [User Settings]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409': # If email is changed to one that's already in use
          description: Email address already in use.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Email address already in use by another account."
        '500':
          $ref: '#/components/responses/ServerError'
...
  /me/smtp/test-connection:
    post:
      summary: Test SMTP connection settings for the current user
      tags: [User Settings, SMTP]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - smtpHost
                - smtpPort
                # smtpSendFrom is good practice for a test
              properties:
                smtpHost:
                  type: string
                  example: "smtp.example.com"
                smtpPort:
                  type: integer
                  example: 587
                smtpEncryption:
                  type: string
                  enum: [none, ssl, starttls]
                  nullable: true
                  example: "starttls"
                smtpUsername:
                  type: string
                  nullable: true
                  example: "user@example.com"
                smtpPassword:
                  type: string
                  format: password
                  nullable: true
                smtpSendFrom:
                  type: string
                  format: email
                  example: "noreply@example.com"
      responses:
        '200':
          description: SMTP connection test result (simulated)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                example:
                  success: true
                  message: "SMTP connection test to smtp.example.com:587 was successful (simulated)."
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'


  # Admin User Management Endpoints
...